strictly depend on sphinx's minor version
Breathe closely integrates with many of Sphinx's internals, especially
integration with Sphinx's C and C++ domains is rather fragile. Changes
often need to be made in both Breathe and Sphinx at the same time.

By pinning the Sphinx minor version in the dependency list it is ensured
only versions that have been (mostly) verified to be compatible with
each other are used, which prevents unexpected breakage. This is often
the case if only Breathe's or only Sphinx's version is pinned in a
project's dependency list.