Fix crash when visiting typedef
When using the @typedef <name> doxygen command, the generated XML does
not have a proper definition, "using ..." or "typedef ..."; this used
to work with versions <= 4.16.0 by assuming everything as a typedef. To
fix the issue, check if "using " is found, otherwise fall back to
considering that "typedef " was used. If not "typedef " is found in the
definition, print a warning.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>