Fix wrong refid when SEPARATE_MEMBER_PAGES is YES
When Doxygen is configured the SEPARATE_MEMBER_PAGES set to YES, the
following Doxygen input:

```
/**
 * This struct must be used with f()
 */
struct t {
  int x;
}

/**
 * @param tx A struct t pointer
 */
void f(struct t *tx) {
  (void)tx;
}
```

The `f()` in struct t's comment generates a <ref> element with `refid`
equal to `some_prefix_compoundid_anchorid`, wheres the `id` in the
refered element ends up with an `id` equal to `some_prefix_anchorid`.
The anchorid here is the compoundid prefix by "_1", with one or two
extra 'g' prefixes for groups, so the `refid` actually has this
information duplicated.

This is a band-aid solution that, when configured enabled in conf.py,
tries to detect the issue in the refids and removes the extra baggage,
so resulting links work.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>