Fix Read the Docs build
The python version is to try to make sure it uses at least 3.5 as that
is what the setup.py file says should work.

We also try to get it to install the packages as required in the
setup.py file. I am very out of date with Python packaging though.