Fix anonymous struct/union usage in C domain
Sphinx's C domain parser expects '.' to be used to separate nested
names. This is currently broken when there are anonymous structs/unions
in the source, where the Doxygen XML provides some nested name with
"::@" in it, which  fails to be parsed properly. This commit adds a
check and does the proper substitution before sending data back to the
domain parser.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>