Fix anchors on pages generated by Doxygen>=1.8.17
Doxygen has changed the structure of generated XML slightly since
version 1.8.13 for "page" generation. Before it would put the anchor
inside a `<varlistentry><term>` block, and on version 1.8.17 and newer,
it inserts the anchor inside the matching `<listitem><para>`.

The `<variablelist>` block structure is still the same, so this commit
only adds support for parsing anchors inside a `<para>` so it works on
both XML generated by old and new versions of Doxygen.

Signed-off-by: Fabio Utzig <fabio.utzig@nordicsemi.no>