Add table parsing fixup to merge row types
"rowspan" is unable to span across different "thead", and "tbody"
elements which make it not useful with the standard row parsing
routines. This commit adds a final step when parsing a table to join all
"thead" and "tbody" rows into just two set of elements.

This still has the limitation that a rowspan from a header won't
span to body elements, but this is not the most usual situation.

Signed-off-by: Fabio Utzig <utzig@apache.org>