Split changelog out & link all PR references
I like the changelog in the README in some ways but it is getting quite
long and large parts of it are of historical interest only so it is a
bit of a burden for the main README.

We link all the PR references with the script below. I was going in
include it but I think it'll be easy enough to copy & paste from
previous lines in the future.

Made with:

    import re

    # Read the changelog
    changelog = open('CHANGELOG.rst').read();

    lines = changelog.split('\n')
    output = []

    # Replace all non-linked end-of-line PR references with a linked version
    for line in lines:
        output.append(
            re.sub(
                '(?P<prefix>[^`])#(?P<number>\d+)$',
                '\g<prefix>`#\g<number> <https://github.com/michaeljones/breathe/pull/\g<number>>`__',
                line, 0, re.M
                )
            )

    # Write the changelog
    open('CHANGELOG.rst', 'w').write("\n".join(output))s