Update project list
Projects that use Breathe based on the GitHub dependents API. Found
with:

  ghtopdep --minstar 100 --rows 50 https://github.com/michaeljones/breathe

From: https://github.com/github-tooling/ghtopdep

There are many more. This is only the top few and the most recognisable.
Some are false positives that I've removed from the list like projects
that vendor their dependencies and a dependency uses Breathe.