rtd: move pip dependencies to conda environment
When using conda the requirements in the main RTD yaml are ignored[1] as
it should be managed by conda. Conda docs[2] indicate how to use pip in
a conda environment, so migrate to that.

Commit e280ba0976f25bbff1a065d8c4773b908b2c2160 earlier introduced conda
to the RTD build system, but migration of environments was not noticed
at the time. With new furo theme dependency changes the build fails as
those dependencies are not installed for the RTD build.

[1] https://docs.readthedocs.io/en/latest/config-file/v2.html#requirements-file
[2] https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#using-pip-in-an-environment